+++ src/spec/pages/Dashboard.spec.tsx
@@ -6,6 +6,7 @@ import Dashboard from '@pages/Dashboard';
 import Alignment from '@components/Alignment';
 import Button from '@components/Button';
 import Graph from '@components/Graph';
+import Input from '@components/Input';
 import PieChart from '@components/PieChart';
 import Section from '@components/Section';
 import Tile from '@components/Tile';
@@ -37,6 +38,11 @@ describe('Dashboard', () => {
 
     it('should contain one Graph component', () => {
       const children = dashboard.find(Graph);
+      expect(children.length).toEqual(1);
+    });
+
+    it('should contain one Input component', () => {
+      const children = dashboard.find(Input);
       expect(children.length).toEqual(1);
     });
 
@@ -87,6 +93,17 @@ describe('Dashboard', () => {
 
       it('should have two records when the Timer stop Button is clicked', () => {
         const wrapper = mount(<Dashboard />);
+        act(() => wrapper.find(Timer).find(Button).at(1).props().handleClick());
+        wrapper.update();
+        expect(wrapper.find(Timesheet).find(Timesheet.Record).length).toBe(2);
+      });
+
+      it('should stack records when timer is stopped and new record name already exists', () => {
+        const wrapper = mount(<Dashboard />);
+        wrapper.find('input').simulate('change', { target: { value: 'test' } });
+        act(() => wrapper.find(Timer).find(Button).at(1).props().handleClick());
+        wrapper.update();
+        wrapper.find('input').simulate('change', { target: { value: 'test' } });
         act(() => wrapper.find(Timer).find(Button).at(1).props().handleClick());
         wrapper.update();
         expect(wrapper.find(Timesheet).find(Timesheet.Record).length).toBe(2);